requirements.txt
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.2.7
django-coreapi==2.3.0
django-cors-headers==3.2.0
djangorestframework==3.10.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
jwt==0.6.1
MarkupSafe==1.1.1
mysqlclient==1.4.5
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
